refactor : service interface 생성 및 불필요한 파일들 삭제
