+++ README.md
@@ -13,3 +13,8 @@
     - [x] wishlist CRUD 테스트
 
 - [x] 전체적인 리팩토링 
+  - [x] 멤버 서비스 인터페이스 생성
+  - [ ] Deprecated API 최신형 API로 변경
+  - [ ] 2주차 PR 리뷰 반영
+  - [ ] MemberId를 UUID로 구현
+  - [ ] @OneToMany, @ManyToOne으로 연관관계 매핑 리팩토링
