--- src/main/java/gift/Login/model/LoginRequest.java
@@ -1,31 +1,0 @@
-package gift.Login.model;
-
-public class LoginRequest {
-    private String email;
-    private String password;
-
-    // Constructors, Getters and Setters
-    public LoginRequest() {
-    }
-
-    public LoginRequest(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
