+++ src/main/resources/application.yml
@@ -1,14 +1,14 @@
 spring:
   application:
     name: spring-gift
-  datasource:
-    url: jdbc:h2:mem:testdb;MODE=MYSQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE # H2 데이터베이스 URL (MySQL 모드 설정 추가)
-    driver-class-name: org.h2.Driver # H2 데이터베이스 드라이버
-    username: sa
   h2:
     console:
       enabled: true # H2 console 활성화 여부
       path: /h2-console
+  datasource:
+    url: jdbc:h2:mem:testdb;MODE=MYSQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE # H2 데이터베이스 URL (MySQL 모드 설정 추가)
+    driver-class-name: org.h2.Driver # H2 데이터베이스 드라이버
+    username: sa
   jpa:
     hibernate:
       ddl-auto: update # 스키마 자동 업데이트
